requirements.txt
# LLM & Orchestration
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-community==0.0.20

# Knowledge Management
chromadb==0.4.22
sentence-transformers==2.3.1

# Web Search & Scraping
duckduckgo-search==4.1.1
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.1

# CLI & UI
click==8.1.7
rich==13.7.0
pyyaml==6.0.1

# Async & Utils
asyncio
python-dotenv==1.0.0
pydantic==2.5.3
